requirements.txt
acres==0.2.0
aiofiles==24.1.0
amqp==5.3.1
annotated-types==0.7.0
anyio==4.8.0
APScheduler==3.11.0
asgiref==3.8.1
async-timeout==5.0.1
beautifulsoup4==4.13.5
billiard==4.2.1
cachetools==5.5.2
celery==5.4.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
ci-info==0.3.0
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
configobj==5.0.9
configparser==7.1.0
contourpy==1.3.1
cron-descriptor==1.4.5
cryptography==44.0.0
cycler==0.12.1
Django==5.1.4
django-background-tasks==1.2.8
django-celery-beat==2.7.0
django-celery-results==2.5.1
django-cors-headers==4.6.0
django-dbshell-plus==1.0.1
django-restframework==0.0.1
django-timezone-field==7.1
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
et_xmlfile==2.0.0
etelemetry==0.3.1
exceptiongroup==1.2.2
filelock==3.17.0
fitz==0.0.1.dev2
fonttools==4.56.0
frontend==0.0.3
google==3.0.0
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.181.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
grpcio==1.74.0
grpcio-status==1.71.2
h11==0.14.0
httplib2==0.22.0
idna==3.10
importlib_resources==6.5.2
isodate==0.6.1
itsdangerous==2.2.0
kiwisolver==1.4.8
kombu==5.5.0
looseversion==1.3.0
lxml==5.3.1
matplotlib==3.10.1
networkx==3.4.2
nibabel==5.3.2
nipype==1.9.2
numpy==2.2.2
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
pathlib==1.0.1
pdfminer.six==20231228
pdfplumber==0.11.5
pillow==11.1.0
prompt_toolkit==3.0.50
proto-plus==1.26.1
protobuf==5.29.5
prov==2.0.1
psycopg2-binary==2.9.10
puremagic==1.28
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pydot==3.0.4
PyJWT==2.10.1
pyparsing==3.2.1
PyPDF2==3.0.1
pypdfium2==4.30.1
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-decouple==3.8
pytils==0.4.1
pytz==2024.2
pyxnat==1.6.3
rdflib==6.3.2
redis==5.2.1
requests==2.32.3
rsa==4.9.1
scipy==1.15.2
simplejson==3.20.1
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
sqlparse==0.5.3
starlette==0.45.3
tools==0.1.9
tqdm==4.67.1
traits==7.0.2
typing-inspection==0.4.1
typing_extensions==4.12.2
tzdata==2025.1
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.3.0
uvicorn==0.34.0
vine==5.1.0
wcwidth==0.2.13
xlrd==2.0.1
XlsxWriter==3.2.2
